backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-cloud-firestore
python-dotenv
rdkit

fastapi
uvicorn[standard]
httpx
pytest

fastapi
uvicorn[standard]
httpx
pytest
